16_Production_RAG_and_Guardrails/pyproject.toml
[project]
name = "16-llmops"
version = "0.1.0"
description = "Production LLMOps and Guardrails with LangGraph Agent Integration"
readme = "README.md"
requires-python = "==3.11.13"
dependencies = [
    "langchain_community==0.3.0",
    "langchain_openai==0.2.0",
    "langchain-core==0.3.59",
    "langgraph>=0.2.0",
    "numpy==1.26.4",
    "tenacity==8.1.0",
    "pymupdf==1.24.10",
    "qdrant-client==1.11.2",
    "langchain_qdrant==0.1.4",
    "langsmith==0.1.125",
    "langchain_huggingface==0.2.0",
    "jupyter>=1.1.1",
    "tavily-python>=0.3.0",
    "arxiv>=2.1.0",
    "cryptography==44.0.3",
    "guardrails-ai==0.5.14",
]